fix: Исправление openAPI файла
Поменян тип с float на double

+++ shared/pkg/openapi/order/v1/oas_json_gen.go
@@ -1165,7 +1165,7 @@ func (s *OrderDto) encodeFields(e *jx.Encoder) {
 	}
 	{
 		e.FieldStart("total_price")
-		e.Float32(s.TotalPrice)
+		e.Float64(s.TotalPrice)
 	}
 	{
 		e.FieldStart("transaction_uuid")
@@ -1247,8 +1247,8 @@ func (s *OrderDto) Decode(d *jx.Decoder) error {
 		case "total_price":
 			requiredBitSet[0] |= 1 << 3
 			if err := func() error {
-				v, err := d.Float32()
-				s.TotalPrice = float32(v)
+				v, err := d.Float64()
+				s.TotalPrice = float64(v)
 				if err != nil {
 					return err
 				}
